Updates/tweaks to our smoke test framework.

+++ test/smoke/types/statechain.go
@@ -1,14 +1,14 @@
 package types
 
 type Pools []struct {
-	BalanceRune  float64 `json:"balance_rune,string"`
-	BalanceAsset float64 `json:"balance_asset,string"`
+	BalanceRune  int64 `json:"balance_rune,string"`
+	BalanceAsset int64 `json:"balance_asset,string"`
 	Asset        struct {
 		Chain  string `json:"chain"`
 		Symbol string `json:"symbol"`
 		Ticker string `json:"string"`
 	} `json:"asset"`
-	PoolUnits   float64 `json:"pool_units,string"`
+	PoolUnits   int64 `json:"pool_units,string"`
 	PoolAddress string  `json:"pool_address"`
 	Status      string  `json:"status"`
 }
@@ -21,6 +21,6 @@ type Staker struct {
 			Symbol string `json:"symbol"`
 			Ticker string `json:"string"`
 		} `json:"asset"`
-		Units float64 `json:"units,string"`
+		Units int64 `json:"units,string"`
 	} `json:"pool_and_units"`
 }
